embedding-service/requirements.txt
fastapi
uvicorn
httpx
langchain
langchain-community
faiss-cpu
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers
einops